+++ glean_parser/lint.py
@@ -300,6 +300,21 @@ def check_metric_on_events_lifetime(
         )
 
 
+def check_event_on_non_events_ping(
+    metric: metrics.Metric, parser_config: Dict[str, Any]
+) -> LintGenerator:
+    """
+    An event metric should usually go on the `events` ping or a custom ping,
+    not on a builtin ping.
+    """
+    disallowed_pings = set(pings.RESERVED_PING_NAMES) - {"default", "events"} | {"health"}
+    if metric.type == "event" and any([ping in disallowed_pings for ping in metric.send_in_pings]):
+        yield (
+            "An event metric should usually go on the `events` ping or a custom ping, "
+            + "not on a builtin ping."
+        )
+
+
 def check_unexpected_unit(
     metric: metrics.Metric, parser_config: Dict[str, Any]
 ) -> LintGenerator:
@@ -466,6 +481,7 @@ METRIC_CHECKS: Dict[
     "EXPIRED": (check_expired_metric, CheckType.warning),
     "OLD_EVENT_API": (check_old_event_api, CheckType.warning),
     "METRIC_ON_EVENTS_LIFETIME": (check_metric_on_events_lifetime, CheckType.error),
+    "EVENT_ON_NON_EVENTS_PING": (check_event_on_non_events_ping, CheckType.warning),
     "UNEXPECTED_UNIT": (check_unexpected_unit, CheckType.warning),
     "EMPTY_DATAREVIEW": (check_empty_datareview, CheckType.warning),
     "HIGHER_DATA_SENSITIVITY_REQUIRED": (